backend-pathway/requirements-pathway.txt
pathway>=0.13.0
openai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
pydantic>=2.5.0
litellm>=1.0.0
tiktoken>=0.5.0
